+++ NewUI/Config/Debugger/SpriteViewerConfig.cs
@@ -15,7 +15,7 @@ namespace Mesen.Config
 		[Reactive] public bool ShowListView { get; set; } = false;
 		[Reactive] public double ListViewHeight { get; set; } = 100;
 
-		[Reactive] public int ImageScale { get; set; } = 2;
+		[Reactive] public double ImageScale { get; set; } = 2;
 		[Reactive] public RefreshTimingConfig RefreshTiming { get; set; } = new();
 
 		public SpriteViewerConfig()
