+++ NewUI/Config/Debugger/TileEditorConfig.cs
@@ -5,7 +5,7 @@ namespace Mesen.Config
 {
 	public class TileEditorConfig : BaseWindowConfig<TileEditorConfig>
 	{
-		[Reactive] public int ImageScale { get; set; } = 8;
+		[Reactive] public double ImageScale { get; set; } = 8;
 		[Reactive] public TileBackground Background { get; set; } = TileBackground.Transparent;
 	}
 }
